Fix local storage issues
