+++ src/modules/taskListClass.js
@@ -1,3 +1,4 @@
+import { checkboxValidation, setLocalStorage } from './interactiveFunctions.js';
 import Task from './taksClass.js';
 
 export default class TaskList {
@@ -6,9 +7,12 @@ export default class TaskList {
   }
 
   addTask(task) {
-    this.Tasks.push(new Task(task, this.Tasks.length));
-    localStorage.setItem('taskList', JSON.stringify(this.Tasks));
+    this.Tasks.push(new Task(task, undefined, this.Tasks.length));
+    setLocalStorage('taskList', this.Tasks);
+    console.log(this.Tasks);
+
     this.showListElements();
+    checkboxValidation(this.Tasks);
   }
 
   deleteTask(ids) {
@@ -21,13 +25,13 @@ export default class TaskList {
     const checkboxStatus = [];
     this.Tasks = [];
     tasksBackUp.forEach((task) => {
-      this.Tasks.push(new Task(task.description, this.Tasks.length));
+      this.Tasks.push(new Task(task.description, task.completed, this.Tasks.length));
       checkboxStatus.push(task.completed);
     });
 
     // Set local storage
-    localStorage.setItem('taskList', JSON.stringify(this.Tasks));
-    localStorage.setItem('checkboxStatus', JSON.stringify(checkboxStatus));
+    setLocalStorage('taskList', this.Tasks);
+    setLocalStorage('checkboxStatus', checkboxStatus);
 
     this.showListElements();
   }
@@ -96,56 +100,7 @@ export default class TaskList {
     });
 
     // Checkboxes event listerner
-    const checkboxInputs = document.querySelectorAll('input[type="checkbox"]');
-    let completedTasks = [];
-
-    // Setting local storage completed tasks if exist
-    if (localStorage.getItem('checkboxStatus')) {
-      const checkboxStatus = JSON.parse(localStorage.getItem('checkboxStatus'));
-
-      for (let index = 0; index < checkboxStatus.length; index += 1) {
-        if (checkboxStatus[index] === true) completedTasks.push(index);
-      }
-    }
-    checkboxInputs.forEach((checkboxInput) => {
-      checkboxInput.addEventListener('change', () => {
-        const taskRow = checkboxInput.parentNode.parentNode;
-        const ellipsisIcon = taskRow.querySelector('.fa-ellipsis-v');
-        const trashIcon = taskRow.querySelector('.fa-trash');
-        const taskDescription = taskRow.querySelector('.task-info label');
-
-        const checkboxStatus = [];
-        if (checkboxInput.checked) {
-          // Show trash icon and add task to completed tasks array
-          ellipsisIcon.classList.add('hide');
-          trashIcon.classList.remove('hide');
-          taskDescription.classList.add('completed-task-text');
-          completedTasks.push(Number(taskRow.id));
-
-          // Set checkbox status true
-          this.Tasks[Number(taskRow.id)].completed = true;
-          this.Tasks.forEach((task) => {
-            checkboxStatus.push(task.completed);
-          });
-          localStorage.setItem('checkboxStatus', JSON.stringify(checkboxStatus));
-        } else {
-          // Show ellipsis icon and remove task from completed tasks array
-          ellipsisIcon.classList.remove('hide');
-          trashIcon.classList.add('hide');
-          taskDescription.classList.remove('completed-task-text');
-          completedTasks = completedTasks.filter((id) => Number(taskRow.id) !== id);
-
-          // Set checkbox status false
-          this.Tasks[Number(taskRow.id)].completed = false;
-          this.Tasks.forEach((task) => {
-            checkboxStatus.push(task.completed);
-          });
-          localStorage.setItem('checkboxStatus', JSON.stringify(checkboxStatus));
-        }
-
-        // Set checkbox status on localstorage
-      });
-    });
+    let completedTasks = checkboxValidation(this.Tasks);
 
     // Form event listener
     const taskListForm = document.querySelector('.tasks-list-form');
@@ -158,6 +113,7 @@ export default class TaskList {
     // Clear completed tasks event listener
     const clearCompletedButton = document.querySelector('.clear-completed-button');
     clearCompletedButton.addEventListener('click', () => {
+      completedTasks = checkboxValidation(this.Tasks);
       this.deleteTask(completedTasks);
     });
 
@@ -180,6 +136,7 @@ export default class TaskList {
 
         taskDescriptionInput.addEventListener('keyup', () => {
           this.Tasks[taskId].description = taskDescriptionInput.value;
+          setLocalStorage('taskList', this.Tasks);
         });
 
         description.classList.add('hide');
